PACT3.0/requirements.txt
# Core API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# LangGraph and AI
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# Data Processing and Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# File Processing
PyPDF2>=3.0.1
python-docx>=1.1.0

# Environment Variables
python-dotenv>=1.0.0

# WebSocket Support
websockets>=12.0

# Async Support
asyncio-mqtt>=0.15.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# PDF Generation
reportlab>=4.0.0
Pillow>=10.0.0
weasyprint>=60.0

# Optional: Production deployment
gunicorn>=21.2.0